+++ src/lbm/D1Q3.hpp
@@ -2,34 +2,68 @@
 
 #include <array>
 #include <concepts>
+#include <cstddef>
 
 #include "LatticeModelBase.hpp"
-
-constexpr std::size_t D1Q3Dimension = 1;
-constexpr std::size_t D1Q3Size = 3;
 
 /**
  * @class D1Q3
  * @brief Class template representing the D1Q3 lattice model.
  *
- * This class template contains the data specific to the D1Q3 lattice model.
- * Accessors are implemented in the CRTP base class LatticeModelBase.
+ * This class template contains the data specific to the `D1Q3` lattice model.
+ * It provides `static constexpr` methods to access the data.
+ * Use the concept `LatticeModel` to restrict template parameters to valid lattice models.
  *
- * @tparam Real The floating-point type.
+ * @tparam T The floating-point type.
  */
-template <std::floating_point Real>
-class D1Q3 : public LatticeModelBase<D1Q3<Real>, Real, D1Q3Dimension, D1Q3Size>
+template <std::floating_point T = double>
+class D1Q3 : public LatticeModelBase
 {
-public:
-    friend class LatticeModelBase<D1Q3<Real>, Real, D1Q3Dimension, D1Q3Size>;
-
 private:
-    static constexpr std::array<std::array<int, D1Q3Dimension>, D1Q3Size> velocities_{
+    static constexpr int dimension_{1};
+
+    static constexpr int size_{3};
+
+    static constexpr std::array<std::array<int, dimension_>, size_> velocities_{
         {{{0}}, {{1}}, {{-1}}}
     };
 
-    static constexpr std::array<Real, D1Q3Size> weights_{
-        static_cast<Real>(4) / static_cast<Real>(6), static_cast<Real>(1) / static_cast<Real>(6),
-        static_cast<Real>(1) / static_cast<Real>(6)
+    static constexpr std::array<T, size_> weights_{
+        static_cast<T>(4) / static_cast<T>(6), static_cast<T>(1) / static_cast<T>(6),
+        static_cast<T>(1) / static_cast<T>(6)
     };
+
+public:
+    using Real = T;
+
+    /**
+     * @brief Returns the spatial dimension D1Q3 (i.e. 1).
+     *
+     * @return The spatial dimension.
+     */
+    static constexpr auto dimension() noexcept -> std::size_t;
+
+    /**
+     * @brief Returns the number of discrete velocity vectors for D1Q3 (i.e. 3).
+     *
+     * @return The number of discrete velocity vectors.
+     */
+    static constexpr auto size() noexcept -> std::size_t;
+
+    /**
+     * @brief Returns the discrete velocity vectors of D1Q3.
+     *
+     * @return The discrete velocity vectors in lattice units.
+     */
+    static constexpr auto
+    velocities() noexcept -> std::array<std::array<int, D1Q3<T>::dimension_>, D1Q3<T>::size_>;
+
+    /**
+     * @brief Returns the quadrature weight for each discrete velocity vector of D1Q3.
+     *
+     * @return The weight for each discrete velocity vector.
+     */
+    static constexpr auto weights() noexcept -> std::array<T, D1Q3<T>::size_>;
 };
+
+#include "D1Q3.tpp"