+++ test/lbm/D3Q27.cpp
@@ -1,5 +1,8 @@
 #include "lbm/D3Q27.hpp"
 
+#include <array>
+#include <cstddef>
+#include <functional>
 #include <numeric>
 
 #include <catch2/catch_approx.hpp>
@@ -11,6 +14,12 @@ SCENARIO("D3Q27 lattice model properties")
     GIVEN("A D3Q27 lattice model")
     {
         using Model = D3Q27<float>;
+
+        std::move_only_function<std::size_t()> model_dimension = &Model::dimension;
+        std::move_only_function<std::size_t()> model_size = &Model::size;
+        std::move_only_function<std::array<std::array<int, Model::dimension()>, Model::size()>()>
+            model_velocities = &Model::velocities;
+        std::move_only_function<std::array<float, Model::size()>()> model_weights = &Model::weights;
 
         const auto expectedDimension{3};
         const auto expectedSize{27};
@@ -39,7 +48,7 @@ SCENARIO("D3Q27 lattice model properties")
 
         WHEN("Getting the dimension")
         {
-            const auto dimension = Model::dimension();
+            const auto dimension = model_dimension();
 
             THEN("The dimension is 3")
             {
@@ -49,7 +58,7 @@ SCENARIO("D3Q27 lattice model properties")
 
         WHEN("Getting the size")
         {
-            const auto size = Model::size();
+            const auto size = model_size();
 
             THEN("The size is 27")
             {
@@ -59,7 +68,7 @@ SCENARIO("D3Q27 lattice model properties")
 
         WHEN("Getting the velocities")
         {
-            const auto velocities = Model::velocities();
+            const auto velocities = model_velocities();
 
             THEN("The 27 velocities are correct")
             {
@@ -83,7 +92,7 @@ SCENARIO("D3Q27 lattice model properties")
 
         WHEN("Getting the weights")
         {
-            const auto weights = Model::weights();
+            const auto weights = model_weights();
 
             THEN("The 27 weights are correct")
             {